src/app/repository-list/repository-list.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RepositoryListComponent } from './repository-list.component';
import {HttpClientTestingModule} from '@angular/common/http/testing';

describe('RepositoryListComponent', () => {
  let component: RepositoryListComponent;
  let fixture: ComponentFixture<RepositoryListComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        HttpClientTestingModule
      ],
      providers: [],
      declarations: [RepositoryListComponent]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(RepositoryListComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should initialise page number as 1', () => {
    expect(component.pageNumber).toBe(1);
  });

  it('should not initialise page number less than 1', () => {
    expect(component.pageNumber).toBeGreaterThanOrEqual(1);
  });
});
